+++ pruebita/src/ctrl/mgrUser.py
@@ -1,3 +1,4 @@
+""" Clase que maneja los Usuarios"""
 from modulo import *
 
 class MgrUser():
@@ -6,21 +7,24 @@ class MgrUser():
         return User.query.all()
     
     def guardar(self, user):
-        db.session.add(user)
-        db.session.commit()
+        if not self.existe(user):
+            db.session.add(user)
+            db.session.commit()
+            return ":guardo usuario:"
+        else:
+            return ":NO guardo usuario:"
         
     def borrar(self, user):
-        if user.estado == "Inactivo":
+        if self.existe(user) and user.estado == "Inactivo" and user.name != "admin":
             db.session.delete(user)
             db.session.commit()
-            return ":borro el usuario:"
+            return ":borro usuario:"
         else:
-            return ":no borro: revisar estado"
+            return ":NO borro usuario:"
             
         
-    def modificar(self, name, passwordNew, confirmacionNew, nombreNew, apellidoNew,
-                    emailNew, telefonoNew, obsNew):
-        user =  self.filtrar(name)
+    def modificar(self, user, passwordNew, confirmacionNew, nombreNew, apellidoNew,
+                    emailNew, telefonoNew, obsNew):       
         if user.estado == "Inactivo":
             user.password = passwordNew
             user.confirmacion = confirmacionNew
@@ -30,9 +34,9 @@ class MgrUser():
             user.telefono = telefonoNew
             user.obs = obsNew        
             db.session.commit()
-            return ":modifico:"
+            return ":modifico usuario:"
         else:
-            return ":no modifico: revisar estado"
+            return ":NO modifico usuario: estado inactivo de usuario" 
         
     def filtrar(self, name):
         return User.query.filter(User.name == name).first_or_404()
@@ -41,37 +45,33 @@ class MgrUser():
     def listarActivo(self):
         return User.query.filter(User.estado == "Activo").all()
     
-    def estado(self, name, estadoNew):
-        user = self.filtrar(name)
-        if estadoNew == "Activo" and self.ceroRol(name) > 0:
+    def estado(self, user, estadoNew):
+        if estadoNew == "Activo" and self.ceroRol(user) > 0:
             user.estado = estadoNew        
             db.session.commit()
-            return ":cambio de estado:"
-        if estadoNew == "Inactivo" and self.ceroRol(name) == 0:
+            return ":cambio estado de usuario:"
+        if estadoNew == "Inactivo" and self.ceroRol(user) == 0:
             user.estado = estadoNew        
             db.session.commit()
-            return ":cambio de estado:"    
-        return ":no cambio de estado:"
+            return ":cambio estado de usuario:"
+        return ": NO cambio estado de usuario:"
     
    
-    def existeRol(self, name, idRol):
+    def existeRol(self, user, idRol):
         rol = Rol.query.filter(Rol.idRol == idRol).first()
-        user = self.filtrar(name)
         if rol in user.roles:
             return True
         else:
             return False
     
-    def ceroRol(self, name):
-        user = self.filtrar(name)
+    def ceroRol(self, user):
         cont = 0
         for rol in user.roles:
             cont = cont + 1
         return cont
         
-    def addRol(self, name, rol):  
-        user = self.filtrar(name)
-        if self.existeRol(name, rol.idRol):
+    def addRol(self, user, rol):  
+        if self.existeRol(user, rol.idRol):
             return ":ya tiene asignado ese rol:"
         else:
             user.roles.append(rol)
@@ -79,9 +79,8 @@ class MgrUser():
             db.session.commit()
             return ":asigno el rol:"
         
-    def removeRol(self, name, rol):   
-        user = self.filtrar(name)
-        if self.existeRol(name, rol.idRol):
+    def removeRol(self, user, rol):   
+        if self.existeRol(user, rol.idRol):
             user.roles.remove(rol)
             db.session.commit()
             if self.ceroRol(name) == 0:
@@ -91,9 +90,25 @@ class MgrUser():
         else:
             return ":no removio el rol:"
         
-    def rolDeUser(self, name, nombreProyecto):
-        user = self.filtrar(name)
+    def rolDeUser(self, user, nombreProyecto):
         for rol in user.roles:
             if rol.ambito == nombreProyecto:
                 return rol
         return None
+
+    def existe(self, user):
+        u = User.query.filter(User.name == user.name).first()
+        if u != None:
+            return True
+        else:
+            return False
+
+    def filtrarXId(self,idUser):
+        return User.query.filter(User.idUser == idUser).first_or_404()
+    
+    def proyectoDeUser(self, user):
+        lista = []        
+        for rol in user.roles:
+            proyecto = Proyecto.query.filter(Proyecto.nombre == rol.ambito).first_or_404()
+            lista.append(proyecto)
+        return lista
